[refacto] Move shared files to a common folder
