+++ api/src/services/auth/user-group/user-group.resolver.ts
@@ -1,7 +1,7 @@
 import { resolve, Resolver } from '@feathersjs/schema'
 import { getValidator } from '@feathersjs/typebox'
 import type { HookContext } from '../../../declarations'
-import { queryValidator } from '../../../schemas/validators'
+import { queryValidator } from '../../../commons/validators'
 import { groupDispatchResolver } from '../group/group.resolver'
 import { userDispatchResolver } from '../user/user.resolver'
 import { UserGroupQuery, userGroupQuerySchema, UserGroupSchema } from './user-group.schema'
