+++ api/src/services/auth/user-group/user-group.schema.ts
@@ -1,6 +1,6 @@
 import { Type, querySyntax, Static, getDataValidator } from '@feathersjs/typebox'
 import { USERGROUP_PROFILE } from '@locokit/definitions'
-import { dataValidator } from '../../../schemas/validators'
+import { dataValidator } from '../../../commons/validators'
 
 // Schema for the basic data model (e.g. creating new entries)
 export const userGroupSchema = Type.Object(
