+++ api/src/services/auth/user/user.schema.ts
@@ -1,6 +1,6 @@
 import { USER_PROFILE } from '@locokit/definitions'
 import { Type, querySyntax, Static } from '@feathersjs/typebox'
-import { dataValidator } from '../../../schemas/validators'
+import { dataValidator } from '../../../commons/validators'
 
 // Schema for the basic data model (e.g. creating new entries)
 export const userSchema = Type.Object(
